djangoapp/requirements.txt
# Framework principal
Django>=5.0.3,<5.1
djangorestframework==3.15.1

# Autenticação JWT
djangorestframework-simplejwt==5.3.1

# Banco de dados PostgreSQL
psycopg2-binary==2.9.9

# Tarefas assíncronas
celery==5.3.6
redis==5.0.3

# Utilidades Django
django-environ==0.11.2   
django-cors-headers==4.3.1
django-filter==24.2
django-jazzmin==3.0.1

requests==2.31.0
reportlab


# Documentação da API
drf-yasg==1.21.7

# Testes
pytest==8.1.1
pytest-django==4.8.0

# Segurança
python-decouple==3.8

